Create helper for building a report Use the full name instead of a user link in reports
