+++ lib/bot/getReportRunner.js
@@ -26,7 +26,7 @@ function runReports(anytimeBot) {
     if(channels.length > 0) {
       log.info('Reporting standups for ' + channels.length + ' channel(s)');
     }
-    
+
     // Iterate over the channels
     un.each(channels, function(channel) {
       // Find standup messages for that channel & for today
@@ -48,50 +48,22 @@ function runReports(anytimeBot) {
           // Iterate over this channels standup messages
           function(callback) {
             un.each(standups, function (standup) {
-              var color = '#000000'.replace(/0/g, function () {
-                return (~~(Math.random()*16)).toString(16);
-              });
-              var fields = [];
-              if (standup.yesterday) {
-                fields.push({
-                    title: 'Yesterday',
-                    value: standup.yesterday,
-                    short: false
-                });
-              }
-              if (standup.today) {
-                fields.push({
-                  title: 'Today',
-                  value: standup.today,
-                  short: false
-                });
-              }
-              if (standup.blockers) {
-                fields.push({
-                  title: 'Blockers',
-                  value: standup.blockers,
-                  short: false
-                });
-              }
-              if (standup.goal) {
-                fields.push({
-                  title: 'Goal',
-                  value: standup.goal,
-                  short: false
-                });
-              }
-              attachments.push({
-                title: '<@'+standup.user+'>',
-                fields: fields,
-                color: color
-              });
+              attachments.push(helpers.getStandupReport(standup));
             });
             callback(null);
           },
           // Send that report off to Slack
           function(callback) {
             report.attachments = attachments;
-            anytimeBot.say(report);
+            anytimeBot.say(report, function(err) {
+              if(err) {
+                console.log('Error sending standup reports:');
+                console.log(err);
+              } else {
+                console.log('Standup reports sent');
+                console.log(report);
+              }
+            });
             callback(null);
           }
         ]);
