+++ lib/helpers/index.js
@@ -2,5 +2,6 @@
 
 module.exports = {
   time: require('./time'),
+  getStandupReport: require('./getStandupReport'),
   getUser: require('./getUser')
 };
